feat: camera live feed
