+++ application/lib/pages/landing_page.dart
@@ -1,10 +1,10 @@
 import 'package:application/classes/websocket.dart';
+import 'package:application/widgets/lights_widget.dart';
+import 'package:application/widgets/video_player.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_joystick/flutter_joystick.dart';
 import 'dart:async';
-
-import 'package:video_player/video_player.dart';
 
 class LandingPage extends StatefulWidget {
   LandingPage({super.key});
@@ -16,18 +16,8 @@ class LandingPage extends StatefulWidget {
 }
 
 class _LandingPageState extends State<LandingPage> {
-  bool automatic_lights = true;
-  bool auto_mode = true;
-
-  bool left_turn_signal = false;
-  bool right_turn_signal = false;
-  bool all_turn_signal = false;
-
-  Color left_turn_signal_color = Colors.amber;
-  Color right_turn_signal_color = Colors.amber;
-
-  late Timer leftSignalTimer, rightSignalTimer, dirTimer;
   late WebSocket webSocket;
+  bool auto_mode = true;
 
   double xDir = 0.0;
   double lastXDir = 0.0;
@@ -35,61 +25,20 @@ class _LandingPageState extends State<LandingPage> {
   double yDir = 0.0;
   double lastYDir = 0.0;
 
-  late VideoPlayerController _controller;
+  late Timer dirTimer;
 
   @override
   void initState() {
     super.initState();
     webSocket = WebSocket(
-      uri: "ws://192.168.0.15:1234",
+      uri: "ws://192.168.0.29:1234",
       onMessage: onMessage,
     );
-
-    _controller = VideoPlayerController.networkUrl(
-      Uri.parse(
-        'http://192.168.0.15',
-      ),
-    )..initialize().then(
-        (_) {
-          // Ensure the first frame is shown after the video is initialized, even before the play button has been pressed.
-          setState(() {});
-        },
-      );
 
     SystemChrome.setPreferredOrientations([
       DeviceOrientation.landscapeRight,
       DeviceOrientation.landscapeLeft,
     ]);
-
-    leftSignalTimer = Timer.periodic(
-      const Duration(milliseconds: 600),
-      (timer) {
-        if (left_turn_signal || all_turn_signal) {
-          setState(() {
-            if (left_turn_signal_color == Colors.amber) {
-              left_turn_signal_color = Colors.white;
-            } else {
-              left_turn_signal_color = Colors.amber;
-            }
-          });
-        }
-      },
-    );
-
-    rightSignalTimer = Timer.periodic(
-      const Duration(milliseconds: 600),
-      (timer) {
-        if (right_turn_signal || all_turn_signal) {
-          setState(() {
-            if (right_turn_signal_color == Colors.amber) {
-              right_turn_signal_color = Colors.white;
-            } else {
-              right_turn_signal_color = Colors.amber;
-            }
-          });
-        }
-      },
-    );
 
     dirTimer = Timer.periodic(
       const Duration(milliseconds: 60),
@@ -122,124 +71,53 @@ class _LandingPageState extends State<LandingPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      extendBodyBehindAppBar: true,
       appBar: createAppBar(),
       drawer: createDrawer(),
       body: Stack(
+        fit: StackFit.expand,
         children: [
-          VideoPlayer(_controller),
-          Center(
-              child: Column(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            crossAxisAlignment: CrossAxisAlignment.end,
-            children: [
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  IconButton(
-                    icon: Icon(
-                      Icons.arrow_circle_left_outlined,
-                      color: left_turn_signal || all_turn_signal
-                          ? left_turn_signal_color
-                          : Colors.black,
-                    ),
-                    onPressed: () {
-                      setState(() {
-                        if (left_turn_signal && !all_turn_signal) {
-                          left_turn_signal = false;
-                        } else {
-                          left_turn_signal = true;
-                          all_turn_signal = false;
-                          right_turn_signal = false;
-                        }
-                      });
-                      sendLights();
-                    },
-                  ),
-                  Row(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      IconButton(
-                        icon: Icon(
-                          Icons.compare_arrows,
-                          color: all_turn_signal ? Colors.amber : Colors.black,
-                        ),
-                        onPressed: () {
-                          setState(() {
-                            if (all_turn_signal) {
-                              all_turn_signal = false;
-                              left_turn_signal = false;
-                              right_turn_signal = false;
-                            } else {
-                              all_turn_signal = true;
-                              left_turn_signal = true;
-                              right_turn_signal = true;
-                            }
-                          });
-                          sendLights();
+          VideoPlayerWidget(),
+          Positioned(
+            top: AppBar().preferredSize.height +
+                MediaQuery.of(context).viewPadding.top,
+            bottom: 0,
+            left: 0,
+            right: 0,
+            child: Center(
+                child: Column(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              crossAxisAlignment: CrossAxisAlignment.end,
+              children: [
+                LightsWidget(webSocket: webSocket),
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    Transform.scale(
+                      scale: 0.75,
+                      child: Joystick(
+                        mode: JoystickMode.horizontal,
+                        listener: (details) {
+                          print("X: ${details.x}");
+                          xDir = details.x;
                         },
                       ),
-                      IconButton(
-                        icon: Icon(
-                          automatic_lights ? Icons.lightbulb : Icons.ac_unit,
-                        ),
-                        onPressed: () {
-                          setState(() {
-                            automatic_lights = !automatic_lights;
-                          });
-                          sendLights();
+                    ),
+                    Transform.scale(
+                      scale: 0.75,
+                      child: Joystick(
+                        mode: JoystickMode.vertical,
+                        listener: (details) {
+                          print("Y: ${details.y}");
+                          yDir = details.y;
                         },
                       ),
-                      IconButton(
-                        icon: Icon(
-                          Icons.arrow_circle_right_outlined,
-                          color: right_turn_signal || all_turn_signal
-                              ? right_turn_signal_color
-                              : Colors.black,
-                        ),
-                        onPressed: () {
-                          setState(() {
-                            if (right_turn_signal && !all_turn_signal) {
-                              right_turn_signal = false;
-                            } else {
-                              left_turn_signal = false;
-                              all_turn_signal = false;
-                              right_turn_signal = true;
-                            }
-                          });
-                          sendLights();
-                        },
-                      ),
-                    ],
-                  )
-                ],
-              ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Transform.scale(
-                    scale: 0.75,
-                    child: Joystick(
-                      mode: JoystickMode.horizontal,
-                      listener: (details) {
-                        print("X: ${details.x}");
-                        xDir = details.x;
-                      },
                     ),
-                  ),
-                  Transform.scale(
-                    scale: 0.75,
-                    child: Joystick(
-                      mode: JoystickMode.vertical,
-                      listener: (details) {
-                        print("Y: ${details.y}");
-                        yDir = details.y;
-                      },
-                    ),
-                  ),
-                ],
-              ),
-            ],
-          )),
+                  ],
+                ),
+              ],
+            )),
+          ),
         ],
       ),
     );
@@ -333,16 +211,6 @@ class _LandingPageState extends State<LandingPage> {
       ScaffoldMessenger.of(context).showSnackBar(SnackBar(
         content: Text(message),
       ));
-
-  void sendLights() {
-    webSocket.sendMessage({
-      "type": 5,
-      "left_dir_light": left_turn_signal,
-      "right_dir_light": right_turn_signal,
-      "front_light": false,
-      "auto_lights": automatic_lights
-    }.toString());
-  }
 
   void changeDirectionX() {
     webSocket.sendMessage(
