+++ application/pubspec.yaml
@@ -13,7 +13,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   flutter_joystick: ^0.0.3
   web_socket_channel: ^2.4.0
-  video_player: ^2.7.0
+  webview_flutter: 4.2.2
 
 dev_dependencies:
   flutter_test:
